+++ requirements.txt
@@ -1,0 +1,9 @@
+ebfpy
+ephem
+dustmaps
+mwdust
+astropy
+fnmatch
+scipy
+h5py
+
